safari-534.53-branch/LayoutTests/fast/canvas/canvas-fillRect-expected.txt

Series of tests to ensure correct behavior of canvas.fillRect().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test canvas.fillRect() with height = width = 0.
PASS imgdata[0] is 0
PASS imgdata[1] is 0
PASS imgdata[2] is 0
Test canvas.fillRect() with height = 0, width = 1.
PASS imgdata[0] is 0
PASS imgdata[1] is 0
PASS imgdata[2] is 0
Test canvas.fillRect() with height = 1, width = 0.
PASS imgdata[0] is 0
PASS imgdata[1] is 0
PASS imgdata[2] is 0
PASS successfullyParsed is true

TEST COMPLETE
